8,1,2 merged, Replace sticker with emoji

+++ submodules/ChatListUI/Sources/ChatListSearchFiltersContainerNode.swift
@@ -2,7 +2,6 @@ import Foundation
 import UIKit
 import AsyncDisplayKit
 import Display
-import SyncCore
 import Postbox
 import TelegramCore
 import TelegramPresentationData
@@ -151,7 +150,7 @@ private final class ItemNode: ASDisplayNode {
 }
 
 enum ChatListSearchFilterEntryId: Hashable {
-    case filter(Int32)
+    case filter(Int64)
 }
 
 enum ChatListSearchFilterEntry: Equatable {
